+++ cpp/open3d/t/geometry/VoxelBlockGrid.h
@@ -263,6 +263,9 @@ private:
 
     // Map: attribute name -> index to access the attribute in SoA.
     std::unordered_map<std::string, int> name_attr_map_;
+
+    // Allocated fragment buffer for reuse in depth estimation
+    core::Tensor fragment_buffer_;
 };
 }  // namespace geometry
 }  // namespace t