+++ cpp/pybind/t/geometry/voxel_block_grid.cpp
@@ -226,8 +226,7 @@ void pybind_voxel_block_grid(py::module& m) {
             "device_id"_a = 0);
 
     vbg.def("save", &VoxelBlockGrid::Save,
-            "Save the voxel block grid to a npz file."
-            "file_name"_a);
+            "Save the voxel block grid to a npz file.", "file_name"_a);
     vbg.def_static("load", &VoxelBlockGrid::Load,
                    "Load a voxel block grid from a npz file.", "file_name"_a);
 }